Fixed little things
