+++ bin/.gitignore
@@ -1,0 +1,8 @@
+/DesafioAnoNascimento01.class
+/DesafioAnoNascimento02.class
+/DesafioContagemRegressiva.class
+/DesafioDespesas.class
+/DesafioMediaIdade.class
+/DesafioNotaAluno.class
+/DesafioSomatoriaIdade.class
+/DesafioTabuada.class
